rename : domain 위치 변경 및 수정, 필요한 라이브러리 적용
