+++ .settings/org.eclipse.wst.common.component
@@ -1,8 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
+        
     <wb-module deploy-name="01.Model2MVCShop(stu)">
+                
         <wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
+                
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/java"/>
+        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/resources"/>
+        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/test/java"/>
+                
         <property name="context-root" value="01.Model2MVCShop(stu)"/>
+                
         <property name="java-output-path" value="/01.Model2MVCShop(stu)/build/classes"/>
+            
     </wb-module>
+    
 </project-modules>
