+++ src/main/java/com/model2/mvc/service/product/dao/ProductDAO.java
@@ -10,8 +10,7 @@ import java.util.Map;
 
 import com.model2.mvc.common.SearchVO;
 import com.model2.mvc.common.util.DBUtil;
-import com.model2.mvc.service.product.domain.ProductVO;
-import com.model2.mvc.service.user.domain.UserVO;
+import com.model2.mvc.service.domain.ProductVO;
 
 public class ProductDAO {
