+++ src/main/java/com/model2/mvc/service/purchase/dao/PurchaseDAO.java
@@ -12,10 +12,10 @@ import java.util.Map;
 import com.model2.mvc.common.SearchVO;
 import com.model2.mvc.common.util.DBUtil;
 import com.model2.mvc.common.util.TranStatusCodeUtil;
-import com.model2.mvc.service.product.domain.ProductVO;
-import com.model2.mvc.service.purchase.domain.PurchaseVO;
+import com.model2.mvc.service.domain.ProductVO;
+import com.model2.mvc.service.domain.PurchaseVO;
+import com.model2.mvc.service.domain.UserVO;
 import com.model2.mvc.service.user.dao.UserDAO;
-import com.model2.mvc.service.user.domain.UserVO;
 
 public class PurchaseDAO {
