+++ src/main/java/com/model2/mvc/service/purchase/impl/PurchaseServiceImpl.java
@@ -5,11 +5,10 @@ import java.util.HashMap;
 import java.util.Map;
 
 import com.model2.mvc.common.SearchVO;
+import com.model2.mvc.service.domain.PurchaseVO;
 import com.model2.mvc.service.product.ProductService;
-import com.model2.mvc.service.product.domain.ProductVO;
 import com.model2.mvc.service.purchase.PurchaseService;
 import com.model2.mvc.service.purchase.dao.PurchaseDAO;
-import com.model2.mvc.service.purchase.domain.PurchaseVO;
 import com.model2.mvc.service.user.dao.UserDAO;
 
 public class PurchaseServiceImpl implements PurchaseService {
