+++ src/main/java/com/model2/mvc/service/user/UserService.java
@@ -3,7 +3,7 @@ package com.model2.mvc.service.user;
 import java.util.Map;
 
 import com.model2.mvc.common.SearchVO;
-import com.model2.mvc.service.user.domain.UserVO;
+import com.model2.mvc.service.domain.UserVO;
 
 
 public interface UserService {
