+++ src/main/java/com/model2/mvc/view/product/GetProductAction.java
@@ -5,8 +5,8 @@ import javax.servlet.http.HttpServletResponse;
 
 import com.model2.mvc.common.util.HistoryUtil;
 import com.model2.mvc.framework.Action;
+import com.model2.mvc.service.domain.ProductVO;
 import com.model2.mvc.service.product.ProductService;
-import com.model2.mvc.service.product.domain.ProductVO;
 import com.model2.mvc.service.product.impl.ProductServiceImpl;
 
 public class GetProductAction extends Action {
