+++ src/main/java/com/model2/mvc/view/purchase/AddPurchaseAction.java
@@ -5,15 +5,15 @@ import javax.servlet.http.HttpServletResponse;
 
 import com.model2.mvc.common.util.TranStatusCode;
 import com.model2.mvc.framework.Action;
+import com.model2.mvc.service.domain.ProductVO;
+import com.model2.mvc.service.domain.PurchaseVO;
+import com.model2.mvc.service.domain.UserVO;
 import com.model2.mvc.service.product.ProductService;
-import com.model2.mvc.service.product.domain.ProductVO;
 import com.model2.mvc.service.product.impl.ProductServiceImpl;
 import com.model2.mvc.service.purchase.PurchaseService;
-import com.model2.mvc.service.purchase.domain.PurchaseVO;
 import com.model2.mvc.service.purchase.impl.PurchaseServiceImpl;
 import com.model2.mvc.service.user.UserService;
 import com.model2.mvc.service.user.dao.UserDAO;
-import com.model2.mvc.service.user.domain.UserVO;
 import com.model2.mvc.service.user.impl.UserServiceImpl;
 
 public class AddPurchaseAction extends Action {
