+++ src/main/java/com/model2/mvc/view/purchase/AddPurchaseViewAction.java
@@ -5,10 +5,10 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import com.model2.mvc.framework.Action;
+import com.model2.mvc.service.domain.ProductVO;
+import com.model2.mvc.service.domain.UserVO;
 import com.model2.mvc.service.product.ProductService;
-import com.model2.mvc.service.product.domain.ProductVO;
 import com.model2.mvc.service.product.impl.ProductServiceImpl;
-import com.model2.mvc.service.user.domain.UserVO;
 
 public class AddPurchaseViewAction extends Action {
