+++ src/main/java/com/model2/mvc/view/purchase/GetPurchaseAction.java
@@ -5,8 +5,8 @@ import javax.servlet.http.HttpServletResponse;
 
 import com.model2.mvc.common.util.PaymentOption;
 import com.model2.mvc.framework.Action;
+import com.model2.mvc.service.domain.PurchaseVO;
 import com.model2.mvc.service.purchase.PurchaseService;
-import com.model2.mvc.service.purchase.domain.PurchaseVO;
 import com.model2.mvc.service.purchase.impl.PurchaseServiceImpl;
 
 public class GetPurchaseAction extends Action {
