+++ src/main/java/com/model2/mvc/view/purchase/ListPurchaseAction.java
@@ -12,10 +12,10 @@ import com.model2.mvc.common.Page;
 import com.model2.mvc.common.SearchVO;
 import com.model2.mvc.common.util.TranStatusCodeUtil;
 import com.model2.mvc.framework.Action;
+import com.model2.mvc.service.domain.PurchaseVO;
+import com.model2.mvc.service.domain.UserVO;
 import com.model2.mvc.service.purchase.PurchaseService;
-import com.model2.mvc.service.purchase.domain.PurchaseVO;
 import com.model2.mvc.service.purchase.impl.PurchaseServiceImpl;
-import com.model2.mvc.service.user.domain.UserVO;
 
 public class ListPurchaseAction extends Action {
