+++ src/main/java/com/model2/mvc/view/user/GetUserAction.java
@@ -4,8 +4,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.model2.mvc.framework.Action;
+import com.model2.mvc.service.domain.UserVO;
 import com.model2.mvc.service.user.UserService;
-import com.model2.mvc.service.user.domain.UserVO;
 import com.model2.mvc.service.user.impl.UserServiceImpl;
