+++ src/main/resources/config/commonservice.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<beans xmlns="http://www.springframework.org/schema/beans"
+
+		xmlns:context = "http://www.springframework.org/schema/context"
+		
+		xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		
+		xsi:schemaLocation=
+						   "http://www.springframework.org/schema/beans 
+							http://www.springframework.org/schema/beans/spring-beans.xsd
+							
+							http://www.springframework.org/schema/context
+							 https://www.springframework.org/schema/context/spring-context.xsd">
+												 
+	<!-- PropertyPlaceholderConfigurer --> 
+	<bean id="propertyConfigurer"
+				class="org.springframework.beans.factory.config.PropertyPlaceholderConfigurer">
+		<property name="location" value="classpath:config/jdbc.properties" />
+	</bean>
+ 	
+	<!--DBCP : Be delegated by mybatis-config.xml -->
+	<bean id="dataSource"	
+				class="org.apache.commons.dbcp.BasicDataSource" >
+		<property name="driverClassName" value="${jdbc.driverClassName}" />
+        <property name="url" value="${jdbc.url}" />
+        <property name="username" value="${jdbc.username}"/>
+        <property name="password" value="${jdbc.password}"/>
+        <property name="initialSize" value="${jdbc.initialSize}"/>
+        <property name="maxActive" value="${jdbc.maxActive}"/>
+	</bean>												 
+												 
+	<!-- SqlSessionFactory -->
+	<bean 	id="sqlSessionFactoryBean" 
+					class="org.mybatis.spring.SqlSessionFactoryBean" >
+		<property name="dataSource" ref="dataSource"/>
+		<property name="configLocation" value="classpath:config/mybatis-config.xml"/>
+	</bean>
+	
+	<!-- SqlSessionTeampate  -->
+	<bean		id="sqlSessionTemplate"		class="org.mybatis.spring.SqlSessionTemplate" >		
+		<constructor-arg ref="sqlSessionFactoryBean"/>
+	</bean>												 
+	
+	<!--  Annotation -->
+	<context:component-scan base-package = "mybatis.service.uservo.impl"/>
+</beans>
+
