+++ src/main/resources/config/jdbc.properties
@@ -1,0 +1,11 @@
+## FileName : jdbc.properties
+jdbc.driverClassName=oracle.jdbc.driver.OracleDriver
+jdbc.url=jdbc:oracle:thin:@127.0.0.1:1521:xe
+jdbc.username=c##scott
+jdbc.password=tiger
+
+jdbc.maxActive=10
+jdbc.initialSize=5
+
+# jdbc.username=c##mybatis
+# jdbc.password=mybatis
