+++ src/main/resources/config/userservice.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<beans xmlns="http://www.springframework.org/schema/beans"
+		xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		xsi:schemaLocation="http://www.springframework.org/schema/beans 
+							http://www.springframework.org/schema/beans/spring-beans.xsd">
+		
+		<bean id = "userDaoImpl" class = "mybatis.service.user.impl.UserDaoImpl">
+			<property name = "sqlSession" ref = "sqlSessionTemplate"/>
+		</bean>										 
+		
+		<bean id = "userServiceImpl" class = "mybatis.service.user.impl.UserServiceImpl">
+			<property name = "userDao" ref = "userDaoImpl"/>
+		</bean>		
+		
+		<bean id = "userVODaoImpl" class = "mybatis.service.uservo.impl.UserVODaoImpl">
+			<property name = "sqlSession" ref = "sqlSessionTemplate"/>
+		</bean>										 
+		
+		<bean id = "userVOServiceImpl" class = "mybatis.service.uservo.impl.UserVOServiceImpl">
+			<property name = "userVODao" ref = "userVODaoImpl"/>
+		</bean>		
+</beans>
