+++ src/main/resources/log4j.properties
@@ -1,0 +1,14 @@
+# root logger definition
+# << sample >> log4j.rootLogger=level,appenderName  
+
+log4j.rootLogger=DEBUG, console
+# log4j.rootLogger=ERROR, console
+
+#root logger appender definition
+# << sample >> log4j.appender.appenderName=appenderClassName(FQCN)
+log4j.appender.console=org.apache.log4j.ConsoleAppender
+
+#root logger layout definition
+# << sample >> log4j.appender.appenderName.layout=layoutClassName(FQCN)
+log4j.appender.console.layout=org.apache.log4j.PatternLayout
+log4j.appender.file.layout.ConversionPattern=%-5p [%t] - %m%n
